Added unit test for sort in updateOne
